Implement basic parallel summation program
This (faulty) program sums from 0 to a constant dividing
the summations across the number of provided processes.
This will serve as a jumping off point for the final submission.
Credits to [Geeks for Geeks](https://www.geeksforgeeks.org/sum-of-an-array-using-mpi/)
for the more accurate and robust program to sum the values in an array.